Release patch 1.2.1 (#73)

+++ src/comps/dataprep/utils/utils.py
@@ -4,10 +4,8 @@
 import os
 import re
 import time
-import unicodedata
 from pathlib import Path
 from typing import List
-import uuid
 from comps.cores.mega.logger import get_opea_logger
 from comps.cores.utils.utils import sanitize_env
 from comps.dataprep.utils.splitter import Splitter
@@ -43,28 +41,14 @@ def save_file_to_local_disk(file: UploadFile) -> str:
     return save_path
 
 
-def save_link_to_local_disk(link_list: List[str]) -> List[str]:
+def save_link_to_local_disk(url: str) -> dict:
     upload_folder = sanitize_env(os.getenv('UPLOAD_PATH', '/tmp/opea_upload'))
     if not os.path.exists(upload_folder):
         Path(upload_folder).mkdir(parents=True, exist_ok=True)
         # Set the directory permission to 700 (owner: rwx)
         os.chmod(upload_folder, 0o700)
-
-    data_collection = parse_html(link_list)  # crawl through the page
-
-    save_paths = []
-    for data, meta in data_collection:
-        doc_id = str(uuid.uuid4()) + ".txt"
-        save_path = Path(os.path.join(upload_folder, doc_id))
-        with save_path.open("w") as fout:
-            try:
-                fout.write(data)
-                fout.close
-                save_paths.append(save_path)
-            except Exception as e:
-                logger.exception(f"Write file failed while presisting links. Exception: {e}")
-                raise
-    return save_paths
+    crawler = Crawler()
+    return crawler.download_file(url, upload_folder) 
 
 
 def parse_files(files: List[UploadFile], splitter: Splitter) -> List[TextDoc]:
@@ -79,6 +63,7 @@ def parse_files(files: List[UploadFile], splitter: Splitter) -> List[TextDoc]:
             logger.info(f"saved file {file.filename} to {saved_path}")
 
             metadata = {
+                'filename': file.filename,
                 'timestamp': time.time()
             }
 
@@ -89,7 +74,7 @@ def parse_files(files: List[UploadFile], splitter: Splitter) -> List[TextDoc]:
             logger.exception(e)
             raise e
         finally:
-            if os.path.exists(saved_path):
+            if saved_path != "" and os.path.exists(saved_path) and not os.path.isdir(saved_path):
                 logger.info(f"Removed {saved_path} after processing")
                 os.remove(saved_path)
 
@@ -101,48 +86,49 @@ def parse_links(links: List[str], splitter: Splitter) -> List[TextDoc]:
 
     for link in links:
         saved_path = ""
+        if not re.match(r"^https?:/{2}\w.+$", link):
+            logger.info(f"The given link/str {link} cannot be parsed.")
+            raise ValueError(f"The given link/str {link} cannot be parsed.")
+
         try:
-            paths = save_link_to_local_disk([link])
-            for path in paths:
-                saved_path = str(path.resolve())
-                logger.info(f"saved link {link} to {saved_path}")
+            # { url, filename, file_path, content_type }
+            parsed_link = save_link_to_local_disk(link)
+            saved_path = parsed_link['file_path']
 
-                metadata = {
-                    'url': link,
-                    'timestamp': time.time()
-                }
+            metadata = {
+                'url': parsed_link['url'],
+                'filename': parsed_link['filename'],
+                'timestamp': time.time()
+            }
 
-                chunks = splitter.split(saved_path)
-                for chunk in chunks:
-                    parsed_texts.append(TextDoc(text=chunk, metadata=metadata))
+             # User parser if file type is html
+            if 'html' in parsed_link['content_type']:
+                logger.info(f"Parsing {parsed_link['file_path']} as HTML")
+                content = load_html_data(parsed_link['file_path'])
+                with open(parsed_link['file_path'], "w", encoding="utf-8") as f:
+                    f.write(content)
+
+            chunks = splitter.split(saved_path)
+            for chunk in chunks:
+                parsed_texts.append(TextDoc(text=chunk, metadata=metadata))
 
         except Exception as e:
             logger.exception(e)
             raise e
         finally:
-            if os.path.exists(saved_path):
+            if saved_path != "" and os.path.exists(saved_path) and not os.path.isdir(saved_path):
                 logger.info(f"Removed {saved_path} after processing")
                 os.remove(saved_path)
 
     return parsed_texts
 
+def load_html_data(file):
+    content = ""
+    with open(file, "r", encoding="utf-8") as f:
+        content = f.read()
 
-def uni_pro(text):
-    """Check if the character is ASCII or falls in the category of non-spacing marks."""
-    normalized_text = unicodedata.normalize("NFKD", text)
-    filtered_text = ""
-    for char in normalized_text:
-        if ord(char) < 128 or unicodedata.category(char) == "Mn":
-            filtered_text += char
-    return filtered_text
-
-
-def load_html_data(url):
     crawler = Crawler()
-    res = crawler.fetch(url)
-    if res is None:
-        return None
-    soup = crawler.parse(res.text)
+    soup = crawler.parse(content)
     all_text = crawler.clean_text(soup.select_one("body").text)
     main_content = ""
     for element_name in ["main", "container"]:
@@ -160,22 +146,5 @@ def load_html_data(url):
     main_content = all_text if main_content == "" else main_content
     main_content = main_content.replace("\n", "")
     main_content = main_content.replace("\n\n", "")
-    main_content = uni_pro(main_content)
     main_content = re.sub(r"\s+", " ", main_content)
     return main_content
-
-
-def parse_html(input):
-    """Parse the uploaded file."""
-    chunks = []
-    for link in input:
-        if re.match(r"^https?:/{2}\w.+$", link):
-            content = load_html_data(link)
-            if content is None:
-                continue
-            chunk = [content.strip(), link]
-            chunks.append(chunk)
-        else:
-            logger.info("The given link/str {} cannot be parsed.".format(link))
-
-    return chunks
